All num buttons are added
All number buttons are added and are functional. Also, coded in the functionality of the percentage button, clear button, and the negative/positive button. Still finding bugs, if any, for the percentage, clear, and negative/positive button. 
For the HTML file, deleted a comment and stored '/100' as the value for the % button.